chore(project): move App into src folder
